Implement email sending. Add email bodies
issue: new password always matches the confirmed password

+++ public/auth.php
@@ -80,7 +80,7 @@ else
         }
         else
         {
-            if(isset($_SESSION[KEY_FORCE_RESET_PASSWORD]) and $_SESSION[KEY_FORCE_RESET_PASSWORD] === true)
+            if(isset($_SESSION[KEY_FORCE_RESET_PASSWORD]) and $_SESSION[KEY_FORCE_RESET_PASSWORD] === true and basename($_SERVER['SCRIPT_NAME']) != 'force_reset_pass.php')
             {
                 echo '<script type="text/javascript"> window.open("' . BASE_URL . 'admin/reset_pass/force_reset_pass.php' . '" , "_self");</script>';
             }
